# client-mac/MacInterviewCopilot/Core/Networking/OpenAPIClient/OpenAPIClient/Classes/OpenAPIs/Models/UserProfile.swift
//
// UserProfile.swift
//
// Generated by openapi-generator
// https://openapi-generator.tech
//

import Foundation
#if canImport(AnyCodable)
import AnyCodable
#endif

public struct UserProfile: Codable, JSONEncodable, Hashable {

    public var id: String?
    public var email: String?
    public var fullName: String?
    public var preferences: UserProfilePreferences?

    public init(id: String? = nil, email: String? = nil, fullName: String? = nil, preferences: UserProfilePreferences? = nil) {
        self.id = id
        self.email = email
        self.fullName = fullName
        self.preferences = preferences
    }

    public enum CodingKeys: String, CodingKey, CaseIterable {
        case id
        case email
        case fullName
        case preferences
    }

    // Encodable protocol methods

    public func encode(to encoder: Encoder) throws {
        var container = encoder.container(keyedBy: CodingKeys.self)
        try container.encodeIfPresent(id, forKey: .id)
        try container.encodeIfPresent(email, forKey: .email)
        try container.encodeIfPresent(fullName, forKey: .fullName)
        try container.encodeIfPresent(preferences, forKey: .preferences)
    }
}


@available(iOS 13, tvOS 13, watchOS 6, macOS 10.15, *)
extension UserProfile: Identifiable {}