# openapi.yaml
openapi: 3.0.3
info:
  title: Mac Interview Copilot API
  version: 1.0.0
  description: >
    Backend API for Mac Interview Copilot (Node/Fastify). 
    Supports SSE (Server-Sent Events) for streaming AI responses.
servers:
  - url: http://localhost:3000/v1
    description: Local Dev Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SignUpRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        fullName: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        user: { $ref: "#/components/schemas/UserProfile" }

    # --- Profile ---
    UserProfile:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        fullName: { type: string }
        preferences:
          type: object
          properties:
            saveSessionHistory: { type: boolean }
            dataRetentionDays: { type: integer }

    # --- Session ---
...
paths:
...
  /auth/signup:
    post:
      summary: Register new user
      tags: [Auth]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignUpRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
  /auth/login:
    post:
      summary: Login user
      tags: [Auth]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }

  # --- Streaming Endpoints (SSE) ---
...